python/src/trocr/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pillow==10.4.0
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
transformers==4.44.2
python-multipart